5_laba/requirements.txt
flask>=3.0,<4
python-docx>=1.1,<2
